+++ .eslintrc.js
@@ -64,7 +64,8 @@ module.exports = {
     'no-useless-constructor': 'warn',
     'comma-dangle': ['error', 'never'],
     'no-param-reassign': 'off',
-    'space-before-function-paren': ["error", "always"]
+    'space-before-function-paren': ["error", "always"],
+    'no-prototype-builtins': 'off'
   },
   'globals': {
     '$': true,