+++ site/news.md
@@ -4,13 +4,13 @@ title: News
 description: News and announcements for all things Bootstrap Table, including new releases.
 ---
 
-## Bootstrap Table 1.15.0
+## Bootstrap Table 1.15.2
 
-<span class="post-date">22 Jun 2019</span>
+<span class="post-date">24 Jun 2019</span>
 
 #### Core
 
-- **New:** Added virtual scroll to support large data sets.
+- **New:** Added `virtualScroll` and `virtualScrollItemHeight` options to support large data.
 - **New:** Added vue component support.
 - **New:** Added support comparisons search(<, >, <=, =<, >=, =>).
 - **New:** Added `detailViewByClick` table option and `detailFormatter` column option.
@@ -75,9 +75,11 @@ description: News and announcements for all things Bootstrap Table, including ne
 - **Update(filter-control):** Fixed loading screen with filter control.
 - **Update(filter-control):** Fixed overwriting the searchText bug.
 - **Update(filter-control):** Fixed filtering does not work json sub-object.
+- **Update(filter-control):** Fixed select filter with formatter.
 - **Update(multiple-sort):** Fixed multiple-sort does not work with data-query-params bug.
 - **Update(page-jump-to):** Fixed `click` bug when paginationVAlign is 'both'.
 - **Update(reorder-columns):** Fixed reorder columns cannot work bug.
+- **Update(reorder-columns):** Fix search and columns bug after reorder columns.
 - **Update(treegrid):** Fixed treegrid cannot work bug.
 
 ## Bootstrap Table 1.14.2